feat: property checker refactoring
