+++ document.lib.rest/Api/TagApiService.cs
@@ -1,5 +1,4 @@
-﻿using System.Globalization;
-using document.lib.shared.Helper;
+﻿using document.lib.shared.Helper;
 
 namespace document.lib.rest.Api;
 
@@ -7,9 +6,9 @@ internal class TagApiService(ITagService tagService)
 {
     public async Task<IResult> GetTagsAsync(GetTagsQueryParameters parameters, HttpContext http)
     {
-        var res = PropertyValidator.Any(parameters, x => x.Id, x => x.Name);
+        var res = PropertyChecker.Values.Any(parameters, x => x.Id, x => x.Name);
         
-        if (PropertyValidator.All(parameters, x => x.Id))
+        if (PropertyChecker.Values.All(parameters, x => x.Id))
         {
             var model = await tagService.GetTagByIdAsync(parameters.Id!.Value);
             return model == null 
@@ -17,7 +16,7 @@ internal class TagApiService(ITagService tagService)
                 : Results.Ok(model);
         }
         
-        if (PropertyValidator.All(parameters, x => x.Name))
+        if (PropertyChecker.Values.All(parameters, x => x.Name))
         {
             var model = await tagService.GetTagByNameAsync(parameters.Name!);
             return model == null 
@@ -25,7 +24,7 @@ internal class TagApiService(ITagService tagService)
                 : Results.Ok(model);
         }
         
-        if (PropertyValidator.All(parameters, x => x.Page, x => x.PageSize))
+        if (PropertyChecker.Values.All(parameters, x => x.Page, x => x.PageSize))
         {
             if (parameters.PageSize!.Value > 50) return Results.BadRequest("Max page size of 50 exceeded");
             var (count, pagedTags) = await tagService.GetTagsPagedAsync(parameters.Page!.Value, parameters.PageSize!.Value);
